+++ src/subscription/websocket.rs
@@ -21,7 +21,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! tears = { version = "0.5", features = ["ws"] }
+//! tears = { version = "0.6", features = ["ws"] }
 //! ```
 //!
 //! ## TLS Support
@@ -36,7 +36,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! tears = { version = "0.5", features = ["ws", "native-tls"] }
+//! tears = { version = "0.6", features = ["ws", "native-tls"] }
 //! ```
 
 use std::hash::{DefaultHasher, Hash, Hasher};